Move and rename files to a more traditional C++ structure
Also, docalign and docjoin are more intuitive than dalign and catdoc 😸🐶
