+++ document-aligner/docjoin.cpp
@@ -1,10 +1,11 @@
-#include "util/file_piece.hh"
-#include "base64.h"
 #include <algorithm>
 #include <vector>
 #include <iostream>
 #include <sstream>
 #include <memory>
+#include "util/file_piece.hh"
+#include "src/base64.h"
+
 
 using namespace bitextor;
 using namespace std;
