+++ document-aligner/src/base64.cpp
