+++ document-aligner/src/base64.h
