+++ document-aligner/src/blocking_queue.h
