+++ document-aligner/src/document.cpp
