+++ document-aligner/src/document.h
