+++ document-aligner/src/murmur_hash.cpp
