+++ document-aligner/src/murmur_hash.h
