+++ document-aligner/src/ngram.cpp
