+++ document-aligner/src/ngram.h
